move video logic to video block
